+++ ma.h
@@ -1,29 +1,25 @@
 // contiene la dimension del espacio original, la matriz y el xo  necesaria para
 // transformar los puntos al sistema de coordenadas original
 
-#include "stdlib.h"
-
-#include "stdio.h" //###
-
 class M_a{
 	private:
-	 int 		 Dim;
-	 int      profundidad;
-	 float	 **Ma;
-	 float    *xa;
+		int 		 Dim;
+		int      profundidad;
+		double	 **Ma;
+		double    *xa;
 
-	 // vect ops
+		// vect ops
 
-	 float    *Mxv(float **M1,float *v);
-	 float    **MxM(float **M1,float **M2);
-	 float    *sum_v(float *v1,float *v2);
+		double    *Mxv(double **M1,double *v);
+		double    **MxM(double **M1,double **M2);
+		double    *sum_v(double *v1,double *v2);
 
 	public:
-	 // constructores
-	 M_a(int d,int p,float **M,float *x);
-	 ~M_a();
+		// constructores
+		M_a(int d,int p,double **M,double *x);
+		~M_a();
 
-	 float *aplicar_Ma_punt(float *punt);
-	 float *aplicar_Ma_vect(float *vect);
-	 M_a   *donar_M_a(float **Mbopt,float *xo); // proporciona el M_a  per posar els punts en les cordenades de l'espai inicial, al nou subspai
+		double *aplicar_Ma_punt(double *punt);
+		double *aplicar_Ma_vect(double *vect);
+		M_a   *donar_M_a(double **Mbopt,double *xo); // proporciona el M_a  per posar els punts en les cordenades de l'espai inicial, al nou subspai
 };
